Shared .py file allows for easier variable management
