+++ gui/shared.py
@@ -1,0 +1,14 @@
+class Shared:
+
+    def __init__(self):
+        self.username = ''
+        self.login = None
+        self.create = None
+        self.main = None
+        self.account = None
+        self.user_events = None
+        self.event_create = None
+        self.event_search = None
+        self.app = None
+        self.sm = None
+        self.nm = None
